[pkg/stanza] Log the entry in `HandleEntryError` when the transformer mode is send_quiet or drop_quiet
The `HandleEntryError` is used in many places for handling parsing
errors. Sometimes the error itself is not very helpful, e.g. the syslog
parser return errors like "parsing error [col 66]", which is quite hard
to understand without seeing the actual entry.

This PR logs the entry when the error mode is `send_quiet` or
`drop_quiet` in debug level. It shouldn't add too much noise
and could be very helpful in debugging.