[pkg/stanza] Log the entry in HandleEntryError function
The `HandleEntryError` is used in many places for handling parsing
errors. Sometimes the error itself is not very helpful, e.g. the syslog
parser return errors like "parsing error [col 66]", which is quite hard
to understand without seeing the actual entry.

Log the entry regardless of the on_error mode for internal usage.