[chore] Refactor the namedpipe input to remove the fsnotify
The current logic is that the input starts 2 goroutines, one for
watching for write events, and another one listens from `watcher.C` and
kick off the process. In other words, it relies on a write event to
start to read from the pipe.

When there is an active writer writing to the pipe, and the collector
exits for whatever reason (or simply can't consume fast enough), the
pipe buffer can be filled and no write can come in. When no write
happens, the watcher doesn't receive any write event, and won't kick off
the process to read the pipe. This is effectively a deadlock, the writer
can't write to the pipe because the pipe buffer is full, and the
collector won't read from the pipe because no write event triggered.

Refactor the namedpipe input to get rid of the watcher. The input will
now keep reading from pipe. If no data or no external writer, it blocks
at the `scan.Scan()` wait for new data or an error when pipe is closed.

Signed-off-by: Mengnan Gong <namco1992@gmail.com>